+++ tests/helpers.php
@@ -97,6 +97,9 @@ if (! function_exists('storage_path')) {
 }
 
 if (! function_exists('fake')) {
+    /**
+     * @return \Faker\Generator
+     */
     function fake(?string $locale = null): mixed
     {
         static $faker = null;